Add skewX and skewY properties to Shape
there's translation, rotation, and scale from matrix transforms but missing skewX and skewY. In order to use these you have to enter manual matrix mode, which is a pain for code maintenance, animations, especially when dealing with groups. This allows skewX and skewY to be set on anything that inherits from Path (so all shapes) for convenience.